# api-contract/openapi.yaml
openapi: 3.1.0
info:
  description: Backend for the Simplicity digital training package. Every endpoint
    except the invitation preview requires a Cognito access token.
  title: Simplicity training API
  version: "1"
servers:
- url: /
security:
- bearerAuth: []
...
paths:
...
  /api/v1/orgs/{orgId}/learning:
    get:
      operationId: listAssignedModules
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignedModuleResponse"
          description: OK
      summary: "Modules assigned to the caller's teams, with their progress"
      tags:
      - Learning
  /api/v1/orgs/{orgId}/learning/media/{assetId}/playback:
    get:
      description: Minted per request after the same assignment check that guards
        the module. Holding an asset id is not authorisation.
      operationId: getPlaybackUrl
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: assetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackResponse"
          description: OK
      summary: A short-lived URL for a video in an assigned module
      tags:
      - Learning
  /api/v1/orgs/{orgId}/learning/sections/{sectionId}/complete:
    put:
      description: Completing the last section completes the module in the same transaction.
      operationId: completeSection
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: sectionId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnerModuleResponse"
          description: OK
      summary: Mark a section as read
      tags:
      - Learning
  /api/v1/orgs/{orgId}/learning/{moduleId}:
    get:
      operationId: readModule
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: moduleId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnerModuleResponse"
          description: OK
      summary: The published version of one assigned module
      tags:
      - Learning
  /api/v1/orgs/{orgId}/learning/{moduleId}/quiz:
    get:
      description: Questions and options only. Which option is correct is never sent
        here.
      operationId: getQuiz
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: moduleId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
          description: OK
      summary: The quiz for an assigned module
      tags:
      - Learning
  /api/v1/orgs/{orgId}/learning/{moduleId}/quiz/attempts:
    post:
      description: "Marked on the server. Returns which questions were right, the\
        \ correct answer, and the author's explanation. Passing completes the module\
        \ if every section is read."
      operationId: submitQuizAttempt
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: moduleId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAttemptRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResultResponse"
          description: OK
      summary: Answer the quiz
      tags:
      - Learning
  /api/v1/orgs/{orgId}/media:
    get:
      operationId: listMedia
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaAssetResponse"
          description: OK
      summary: The organisation's video library
      tags:
      - Media
    post:
      description: Returns a presigned URL the browser PUTs to directly. Video never
        passes through the API on the way in.
      operationId: registerUpload
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUploadRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadTargetResponse"
          description: Created
      summary: Register a video and get somewhere to put it
      tags:
      - Media
  /api/v1/orgs/{orgId}/media/{assetId}:
    delete:
      description: Any section using it keeps its writing and loses the video.
      operationId: deleteMedia
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: assetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      summary: Delete a video
      tags:
      - Media
  /api/v1/orgs/{orgId}/media/{assetId}/uploaded:
    post:
      description: Hands the file to the transcoder; the asset becomes PROCESSING.
      operationId: completeUpload
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: assetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetResponse"
          description: OK
      summary: Report that the upload finished
      tags:
      - Media
...
components:
  schemas:
...
    AnswerInput:
      type: object
      properties:
        optionId:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
      required:
      - optionId
      - questionId
...
    AssignedModuleResponse:
      type: object
      properties:
        completedSectionCount:
          type: integer
          format: int32
        hasQuiz:
          type: boolean
        moduleId:
          type: string
          format: uuid
        quizPassed:
          type: boolean
        sectionCount:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - NOT_STARTED
          - IN_PROGRESS
          - COMPLETED
          - NEEDS_REDOING
        summary:
          type: string
        title:
          type: string
        versionId:
          type: string
          format: uuid
    AttemptResultResponse:
      type: object
      properties:
        attemptNumber:
          type: integer
          format: int32
        correctCount:
          type: integer
          format: int32
        passed:
          type: boolean
        questionCount:
          type: integer
          format: int32
        questions:
          type: array
          items:
            $ref: "#/components/schemas/MarkedQuestion"
...
    LearnerModuleResponse:
      type: object
      properties:
        completedSectionIds:
          type: array
          items:
            type: string
            format: uuid
        hasQuiz:
          type: boolean
        moduleId:
          type: string
          format: uuid
        quizPassed:
          type: boolean
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionResponse"
        status:
          type: string
          enum:
          - NOT_STARTED
          - IN_PROGRESS
          - COMPLETED
          - NEEDS_REDOING
        summary:
          type: string
        title:
          type: string
        versionId:
          type: string
          format: uuid
    MarkedQuestion:
      type: object
      properties:
        chosenOptionId:
          type: string
          format: uuid
        correctOptionId:
          type: string
          format: uuid
        explanation:
          type: string
        prompt:
          type: string
        questionId:
          type: string
          format: uuid
        wasCorrect:
          type: boolean
    MediaAssetResponse:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          format: int32
        failureReason:
          type: string
        filename:
          type: string
        sizeBytes:
          type: integer
          format: int64
        status:
          type: string
...
    PlaybackResponse:
      type: object
      properties:
        expiresInSeconds:
          type: integer
          format: int32
        url:
          type: string
...
    QuizOptionResponse:
      type: object
      properties:
        label:
          type: string
        optionId:
          type: string
          format: uuid
        position:
          type: integer
          format: int32
    QuizQuestionResponse:
      type: object
      properties:
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOptionResponse"
        position:
          type: integer
          format: int32
        prompt:
          type: string
        questionId:
          type: string
          format: uuid
    QuizResponse:
      type: object
      properties:
        attemptCount:
          type: integer
          format: int32
        passed:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionResponse"
        versionId:
          type: string
          format: uuid
    RegisterUploadRequest:
      type: object
      properties:
        contentType:
          type: string
          minLength: 1
        filename:
          type: string
          minLength: 1
        sizeBytes:
          type: integer
          format: int64
          minimum: 0
      required:
      - contentType
      - filename
...
    SectionResponse:
      type: object
      properties:
        body:
          type: string
        mediaAssetId:
          type: string
          format: uuid
        position:
          type: integer
          format: int32
        sectionId:
          type: string
          format: uuid
        title:
          type: string
...
    SubmitAttemptRequest:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
          maxItems: 100
          minItems: 0
      required:
      - answers
...
    UploadTargetResponse:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        uploadUrl:
          type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Cognito access token
      scheme: bearer
      type: http